dashboard sidebar

+++ vue/src/components/PersonalBar.vue
@@ -3,14 +3,19 @@
     <template slot="text">
       {{$t('personal-bar.welcome', {name: user.name})}}
     </template>
-    <b-dropdown-item to="#">Profile</b-dropdown-item>
-    <b-dropdown-item to="#">Signout</b-dropdown-item>
+    <b-dropdown-item :to="{name: 'site.dashboard'}">
+      {{$t('personal-bar.dashboard')}}
+    </b-dropdown-item>
+    <b-dropdown-divider />
+    <b-dropdown-item v-on:click="onSignOut">
+      {{$t('personal-bar.sign-out')}}
+    </b-dropdown-item>
   </b-nav-item-dropdown>
   <b-nav-item-dropdown right v-else="user">
     <template slot="text">
       {{$t('personal-bar.sign-in-or-up')}}
     </template>
-    <b-dropdown-item v-if="l" v-for="l in links" v-bind:key="l" :to="{name: l.href}">
+    <b-dropdown-item v-if="l" v-for="(l, i) in links" :key="i" :to="{name: l.href}">
       {{$t(`${l.href}.title`)}}
     </b-dropdown-item>
     <b-dropdown-divider v-else />
@@ -40,11 +45,13 @@ export default {
   },
   methods: {
     onSignOut () {
-      _delete('/users/sign-out', function (rst) {
-        sessionStorage.clear()
-        this.$store.commit('signOut')
-        this.$router.push({name: 'site.home'})
-      }.bind(this))
+      if (confirm(this.$t('are-you-sure'))) {
+        _delete('/users/sign-out').then(function (rst) {
+          sessionStorage.clear()
+          this.$store.commit('signOut')
+          this.$router.push({name: 'site.home'})
+        }.bind(this)).catch(alert)
+      }
     }
   }
 }
